+++ bin/strata-cli/src/cmd/deposit.rs
@@ -21,7 +21,7 @@ use crate::{
     recovery::DescriptorRecovery,
     seed::Seed,
     settings::Settings,
-    signet::{broadcast_tx, get_fee_rate, log_fee_rate, print_explorer_url, SignetWallet},
+    signet::{log_fee_rate, persist::WalletPersistWrapper, print_explorer_url, SignetWallet},
     strata::StrataWallet,
     taproot::{ExtractP2trPubkey, NotTaprootAddress},
 };
@@ -51,12 +51,12 @@ pub async fn deposit(
     let requested_strata_address =
         strata_address.map(|a| StrataAddress::from_str(&a).expect("bad strata address"));
     let mut l1w =
-        SignetWallet::new(&seed, settings.network, settings.sync_backend.clone()).unwrap();
+        SignetWallet::new(&seed, settings.network, settings.signet_backend.clone()).unwrap();
     let l2w = StrataWallet::new(&seed, &settings.strata_endpoint).unwrap();
 
     l1w.sync().await.unwrap();
     let recovery_address = l1w.reveal_next_address(KeychainKind::External).address;
-    l1w.persist().unwrap();
+    WalletPersistWrapper::persist(&mut l1w).unwrap();
 
     let strata_address = requested_strata_address.unwrap_or(l2w.default_signer_address());
     let _ = term.write_line(&format!(
@@ -101,7 +101,9 @@ pub async fn deposit(
         style(bridge_in_address.to_string()).yellow()
     ));
 
-    let fee_rate = get_fee_rate(fee_rate, settings.sync_backend.clone(), 1)
+    let fee_rate = settings
+        .signet_backend
+        .get_fee_rate(fee_rate, 1)
         .await
         .expect("valid fee rate");
     log_fee_rate(&term, &fee_rate);
@@ -143,7 +145,9 @@ pub async fn deposit(
 
     let pb = ProgressBar::new_spinner().with_message("Broadcasting transaction");
     pb.enable_steady_tick(Duration::from_millis(100));
-    broadcast_tx(&tx, settings.sync_backend.clone())
+    settings
+        .signet_backend
+        .broadcast_tx(&tx)
         .await
         .expect("successful broadcast");
     let txid = tx.compute_txid();
